Solution day 14 part 2

+++ utils.h
@@ -41,8 +41,8 @@ class Solver {
   int Solve_day_13_part1();
   int Solve_day_13_part2();
 
-  int Solve_day_14_part1();
-  int Solve_day_14_part2();
+  long Solve_day_14_part1();
+  long long Solve_day_14_part2();
 };
 
 #endif
